# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/bank":{"post":{"operationId":"BankController_add","summary":"Adds new bank","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BankDTO"}}}},"responses":{"201":{"description":""}},"tags":["Bank"]},"get":{"operationId":"BankController_get","summary":"Fetches bank list","parameters":[],"responses":{"200":{"description":""}},"tags":["Bank"]}},"/api/v1/bank/{id}":{"delete":{"operationId":"BankController_deleteOne","summary":"Deletes a bank","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Bank"]}},"/api/v1/product-category":{"post":{"operationId":"ProductCategoryController_add","summary":"Adds new category","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCategoryDTO"}}}},"responses":{"201":{"description":""}},"tags":["Product Category"]},"get":{"operationId":"ProductCategoryController_get","summary":"Fetches category list","parameters":[],"responses":{"200":{"description":""}},"tags":["Product Category"]}},"/api/v1/product-category/{id}":{"put":{"operationId":"ProductCategoryController_update","summary":"Updates a category","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product Category"]},"delete":{"operationId":"ProductCategoryController_deleteOne","summary":"Deletes a category","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product Category"]}},"/api/v1/vehicle":{"post":{"operationId":"VehicleController_add","summary":"Adds new vehicle","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleDTO"}}}},"responses":{"201":{"description":""}},"tags":["Vehicle"]},"get":{"operationId":"VehicleController_get","summary":"Fetches vehicle list","parameters":[],"responses":{"200":{"description":""}},"tags":["Vehicle"]}},"/api/v1/vehicle/{id}":{"put":{"operationId":"VehicleController_update","summary":"Updates a vehicle","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle"]},"delete":{"operationId":"VehicleController_deleteOne","summary":"Deletes a vehicle","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Vehicle"]}}},"info":{"title":"Storedash Utility Module","description":"Endpoints for all platforms (mobile & web portal)","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"BankDTO":{"type":"object","properties":{"name":{"type":"string","example":"First Bank","description":"Bank Name"},"code":{"type":"string","example":"101","description":"Bank Code"}},"required":["name","code"]},"ProductCategoryDTO":{"type":"object","properties":{"name":{"type":"string","example":"Groceries","description":"Category Name"}},"required":["name"]},"VehicleDTO":{"type":"object","properties":{"brand":{"type":"string","example":"Toyota","description":"Vehicle Name"},"make":{"type":"string","example":"Hilux","description":"Vehicle Brand"}},"required":["brand","make"]}}}}